+++ fsleyes/actions/frameactions.py
@@ -23,7 +23,6 @@ from   fsleyes.frame import FSLeyesFrame
 def addViewPanel(self, vpType, **kwargs):
     """Function shared by the add*Panel functions below. """
     vp = self.addViewPanel(vpType, **kwargs)
-    self.viewPanelDefaultLayout(vp)
     vp.SetFocus()
     return vp
 
@@ -140,7 +139,7 @@ def openHelp(self, *args, **kwargs):
     if op.exists(url):
         webpage.openFile(url)
     else:
-        url = 'https://users.fmrib.ox.ac.uk/~paulmc/fsleyes/userdoc/latest/'
+        url = 'https://open.win.ox.ac.uk/pages/fsl/fsleyes/fsleyes/userdoc/'
         webpage.openPage(url)